Reorder projects architecture
Used multy-solution approach. Now all the solution are in the same folder. The "Speedster" proyect references all the differents modules while the "<name> Prototype" reference only one of this modules plus the one used to test and develop that particular proyect. This allows all the changes done to the module in the prototype solution to be immediately effective in "Speedster"